[dev-frontend-fb] Implement all features left (#164)
* feat: add loading to signup process

* feat: block invalid email at signup

* fix: block 401 error for noti

* fix: fix rerendering issue when changing docspace head

* feat: block unauthorized projects

* feat: allow manager change

* feat: allow only project members in AddTask

* feat: allow task assignee change

* fix: disallow status change for null assignee task

* feat: create profile page

+++ frontend/sprint/src/pages/ProjectManage.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { List, Avatar, Button, Modal, AutoComplete, Input, message } from 'antd';
+import { List, Avatar, Button, Modal, AutoComplete, Input, message, Tag } from 'antd';
 import AutoOption from '@components/AutoOption';
 import { UserType } from '@store/zustand/user';
 import useBindStore from '@store/zustand';
@@ -28,14 +28,20 @@ const ProjectManage: React.FC = () => {
   const selectProject = useBindStore(state => state.selectProject);
   const editProject = useBindStore(state => state.editProject);
   const deleteProject = useBindStore(state => state.deleteProject);
+  const assignManager = useBindStore(state => state.assignManager);
   const isManager = project?.manager === user?.id;
   const [projectName, setProjectName] = useState(project?.name);
   const [projectSubject, setProjectSubject] = useState(project?.subject);
 
   useEffect(() => {
+    setProjectName(project?.name);
+    setProjectSubject(project?.subject);
+  }, [project]);
+
+  useEffect(() => {
     const asyncGetAutoComplete = async (): Promise<void> => {
       if (query === '') return;
-      const autoComplete = await getAutoComplete(query);
+      const autoComplete = await getAutoComplete(query, -1);
       setQueryList(autoComplete);
     };
     void asyncGetAutoComplete();
@@ -67,7 +73,18 @@ const ProjectManage: React.FC = () => {
     await selectProject(parseInt(projectId));
   };
 
+  const onDelegateClick = async (userId: number): Promise<void> => {
+    if (!window.confirm('Do you want to give the manager to this user? Your authorities will be lost.')) return;
+    if (projectId === undefined) return;
+    await assignManager(parseInt(projectId), userId);
+    await selectProject(parseInt(projectId));
+  };
+
   const onLeaveClick = async (userId: number): Promise<void> => {
+    if (isManager) {
+      alert('You should assign a manager before leaving the project');
+      return;
+    }
     if (!window.confirm('Do you want to leave this project?')) return;
     if (projectId === undefined) return;
     await deleteMember(parseInt(projectId), userId);
@@ -138,13 +155,23 @@ const ProjectManage: React.FC = () => {
               renderItem={item => (
                 <List.Item
                   actions={[user?.id !== item.id
-                    ? isManager ? <a key="list-delete" onClick={() => { void onDeleteClick(item.id); }}>delete</a> : null
+                    ? isManager
+                      ? <>
+                          <a className="give-manager" onClick={() => { void onDelegateClick(item.id); }}>delegate</a>
+                          <a key="list-delete" onClick={() => { void onDeleteClick(item.id); }}>delete</a>
+                        </>
+                      : null
                     : <a key="list-delete" onClick={() => { void onLeaveClick(item.id); }}>leave</a>
                   ]}
                 >
                   <List.Item.Meta
                     avatar={<Avatar>{iconString(item.username)}</Avatar>}
-                    title={item.username}
+                    title={
+                    <div>
+                      {item.username}
+                      {item.id === project?.manager ? <Tag className="manager-tag">Manager</Tag> : null}
+                    </div>
+                    }
                     description={item.email}
                   />
                 </List.Item>
